feat: support detecting root package lib name
Prior to this commit the crate name was unconditionally used as the
JSON file name `CargoRustDocJson` would expect to be produced by running
`rustdoc`. In some cases the package under test may specify an
alternative `lib.name` in its `Cargo.toml`, which would result in a file
not found error when using the crate name instead of that alternative
name.

This commit updates the processing logic so that:

1. We use the cargo metadata to find the root package.
2. We iterate the root package looking for all `lib` type targets.
  i. If there is only one lib target, we extract its name to use for the
  `CargoRustDocJson` expected filename.
  ii. If there is more than one lib target, we return an error - this
      shouldn't be possible based on today's `Cargo.toml` semantics.
  iii. If there are no lib targets, we return an error - this indicates
       the crate only has non-lib targets.

With the new logic in place we can remove the existing code for
determining the Crate name to pass forward to `CargoRustDocJson`, it was
only used for the expected JSON file name.