Support configuration via Cargo.toml metadata (#116)
Previously it was only possible to specify configuration (e.g. to allow
certain types) via a separate TOML config file, provided via the
`--config` CLI arg.

In this commit we add support for specifying the same config in the
Crate's `Cargo.toml` in a metadata section like:

```toml
[package.metadata.cargo_check_external_types]
allow_alloc = true
allow_core = true
allow_std = true
allowed_external_types = ["foo*", "bar::BazStruct"]
```

When present in the `Cargo.toml` of the crate under test, this config
will be used automatically. When omitted, the default config will be
used.

If an explicit TOML config file is provided via `--config`, it will take
precedence and override what may be found in the metadata.

A new integration test is added that demonstrates the metadata config in
action, as well as the override behaviour with an explicit TOML config
file.