Bump rustdoc-types to v25
When using `cargo-check-external-types` with the latest nightly release I saw:

```
The version of rustdoc being used produces JSON format version 29, but this tool requires format version 28. This can happen if the locally installed version of rustdoc doesn't match the rustdoc JSON types from the `rustdoc-types` crate.

If this occurs with the latest Rust nightly and the latest version of this tool, then this is a bug, and the tool needs to be upgraded to the latest format version.

Otherwise, you'll need to determine a Rust nightly version that matches this tool's supported format version (or vice versa).
```

This change adds support for the most recent nightly versions using rustdoc format version 29.

The changelog for `rustdoc-types` can be found here https://github.com/aDotInTheVoid/rustdoc-types/blob/trunk/CHANGELOG.md#v0250---2024-04-19